+++ .gitignore
@@ -16,3 +16,4 @@ local.properties
 signing.properties
 /.kotlin
 /library/hook
+*.apk
