+++ app/src/main/cpp/CMakeLists.txt
@@ -3,18 +3,10 @@ project(hyperceiler_home LANGUAGES CXX)
 
 add_library(hyperceiler_home SHARED hyperceiler_home.cpp)
 
-option(HYPERCEILER_DOCK_NATIVE_PROBE "Build the opt-in, bounded OS4 scale diagnostic" OFF)
 option(HYPERCEILER_DOCK_NATIVE_MOTION "Follow pinned OS4 native recents animation" ON)
-if(HYPERCEILER_DOCK_NATIVE_PROBE OR HYPERCEILER_DOCK_NATIVE_MOTION)
-    enable_language(ASM)
-    target_sources(hyperceiler_home PRIVATE dock_native_probe.cpp dock_native_probe_arm64.S)
-    target_compile_definitions(hyperceiler_home PRIVATE HYPERCEILER_DOCK_NATIVE_OBSERVER=1)
-endif()
-if(HYPERCEILER_DOCK_NATIVE_PROBE)
-    target_compile_definitions(hyperceiler_home PRIVATE HYPERCEILER_DOCK_NATIVE_PROBE=1)
-endif()
 if(HYPERCEILER_DOCK_NATIVE_MOTION)
-    target_sources(hyperceiler_home PRIVATE dock_native_motion.cpp dock_native_motion_arm64.S)
+    enable_language(ASM)
+    target_sources(hyperceiler_home PRIVATE dock_native_hooks.cpp dock_native_motion.cpp dock_native_motion_arm64.S)
     target_compile_definitions(hyperceiler_home PRIVATE HYPERCEILER_DOCK_NATIVE_MOTION=1)
 endif()
